+++ src/LogFileInterface.php
@@ -7,4 +7,8 @@ interface LogFileInterface {
 	function getRealPath():string;
 	function getPath():string;
 	function getFilename():string;
+	function getTimestamp():string;
+	function setTimezone($timezone=null):LogFileInterface;
+	function setTimeFormat(string $format):LogFileInterface;
+	function setOptions(array $options):LogFileInterface;
 }
